feat: move find request into own file
